fix but type of variable

+++ examples/tiny32_pzem003_getParam/tiny32_pzem003_getParam.ino
@@ -23,7 +23,7 @@ tiny32_v3 mcu; //define object
 uint8_t id = 1; //Address if PZEM-003, You can change here if it differance
 
 float volt, amp, power;
-uint16_t energy;
+uint32_t energy;
 
 
 void setup()
